requirements.txt
asgiref==3.9.1
django==4.2.23
djangorestframework==3.16.1
sqlparse==0.5.3
typing-extensions==4.15.0
tzdata==2025.2
